--- src/components/atoms/index.jsx
@@ -1,4 +1,0 @@
-import { CustomActionIcon } from "./CustomActionIcon"
-import { CustomButton } from "./CustomButton"
-
-export {CustomActionIcon, CustomButton }
