+++ src/components/molecules/FFModal/FFAddCategoryModal.jsx
@@ -1,23 +1,24 @@
 import React, { useState, useEffect, useRef } from 'react'
-import { DUMMY_CATEGORIES } from '../../utils/constants'
-import { imageUpload, removeImage } from '../../utils/imagesFunctions'
+
+import { imageUpload, removeImage } from '../../../utils/imagesFunctions'
 import { useDispatch, useSelector } from 'react-redux'
 import {
-  edit_category_modal,
+  add_category_modal,
+  postCategory,
   getAllCategories,
   isSuccess,
   isError,
-  editCategoryModalOpen,
-  updateCategory,
+  addCategoryModal,
   message,
-} from '../../redux/features/categorySlice'
+} from '../../../redux/features/categorySlice'
 
-const EditCategoryModal = () => {
+const FFAddCategoryModal = () => {
   const dispatch = useDispatch()
-  const edit_category_detail_modal = useSelector(edit_category_modal)
+  const add_category_detail_modal = useSelector(add_category_modal)
   const succesStatus = useSelector(isSuccess)
   const errorStatus = useSelector(isError)
   const responseMessage = useSelector(message)
+
   const alert = (msg, type) => (
     <div
       className={`text-${type} flex felx-row  justify-center items-center py-2 px-4 w-full`}
@@ -33,69 +34,63 @@ const EditCategoryModal = () => {
 
   const [imageAdded, setImageAdded] = useState(false)
 
-  const [editFormData, setEditformdata] = useState({
-    _id: null,
+  const [fData, setFdata] = useState({
+    category_id: '',
     category_name: '',
-
-    category_status: '',
+    category_status: 'Available',
     category_description: '',
     category_image: '',
-
     success: false,
     error: false,
   })
   useEffect(() => {
-    setEditformdata({
-      category_name: edit_category_detail_modal.category_name,
-
-      category_status: edit_category_detail_modal.category_status,
-      category_description: edit_category_detail_modal.category_description,
-      category_image: edit_category_detail_modal.category_image,
-    })
-  }, [edit_category_detail_modal])
-
+    console.log(add_category_detail_modal)
+  }, [])
   useEffect(() => {
     setImageAdded(true)
     setSelectedFile(null)
-  }, [editFormData?.product_image])
+    console.log(add_category_detail_modal)
+  }, [fData?.category_image])
 
   const handleSubmit = async (e) => {
     e.preventDefault()
-    if (editFormData?.product_image) {
-      console.log('UploaImage ')
-    } else {
-      console.log('Image uploading')
+
+    if (fData.category_image) {
+      setFdata({ ...fData, error: 'Please upload at least 1 image' })
+      setTimeout(() => {
+        setFdata({ ...fData, error: false })
+      }, 2000)
     }
+
     try {
-      dispatch(
-        updateCategory({
-          category: {
-            category_image: editFormData.category_image,
-            category_name: editFormData.category_name,
-            category_status: editFormData.category_status,
-            category_description: editFormData.category_description,
-          },
-          category_id: edit_category_detail_modal._id,
-        }),
-      )
+      const { success, error, ...data } = fData
+      dispatch(postCategory(data))
 
       if (succesStatus) {
         dispatch(getAllCategories())
-        setEditformdata({ ...editFormData, success: succesStatus })
+        setFdata({
+          category_name: '',
+          category_status: '',
+          category_description: '',
+          category_image: '',
+          success: succesStatus,
+          error: false,
+        })
         setTimeout(() => {
-          return setEditformdata({
-            ...editFormData,
-            success: succesStatus,
+          setFdata({
+            ...fData,
+            category_name: '',
+            category_status: '',
+            category_description: '',
+            category_image: '',
+            success: false,
+            error: false,
           })
         }, 2000)
       } else if (errorStatus) {
-        setEditformdata({ ...editFormData, error: errorStatus })
+        setFdata({ ...fData, success: false, error: errorStatus })
         setTimeout(() => {
-          return setEditformdata({
-            ...editFormData,
-            error: false,
-            success: false,
-          })
+          return setFdata({ ...fData, error: false, success: false })
         }, 2000)
       }
     } catch (error) {
@@ -129,7 +124,7 @@ const EditCategoryModal = () => {
     imageUpload(selectedFile, 'categoryImages')
       .then((imageUrl) => {
         // push the imageUrl to the imageUrl array
-        setEditformdata((prevState) => ({
+        setFdata((prevState) => ({
           ...prevState,
           error: false,
           success: false,
@@ -149,7 +144,7 @@ const EditCategoryModal = () => {
     removeImage(imageUrl)
       .then(
         // Remove the imageUrl from the formData
-        setEditformdata((prevState) => ({
+        setFdata((prevState) => ({
           ...prevState,
           error: false,
           success: false,
@@ -169,10 +164,10 @@ const EditCategoryModal = () => {
       <div
         onClick={() => {
           console.log('background clicked')
-          dispatch(editCategoryModalOpen({ open: false }))
+          dispatch(addCategoryModal(false))
         }}
         className={`${
-          edit_category_detail_modal.modal ? '' : 'hidden'
+          add_category_detail_modal ? '' : 'hidden'
         } fixed top-0 left-0 z-30 w-full h-full bg-black opacity-50`}
       />
       {/* End Black Overlay */}
@@ -180,21 +175,18 @@ const EditCategoryModal = () => {
       {/* Modal Start */}
       <div
         className={`${
-          edit_category_detail_modal.modal ? '' : 'hidden'
+          add_category_detail_modal ? '' : 'hidden'
         } fixed inset-0 flex items-center z-30 justify-center overflow-auto`}
       >
         <div className="mt-4 md:mt-0 relative bg-[#F5F3F0] w-8/12 md:w-3/6 shadow-lg flex flex-col items-center space-y-4 px-4 py-4 md:px-8 rounded-xl">
           <div className="flex items-center justify-between w-full pt-4">
             <span className="text-left font-semibold text-2xl tracking-wider">
-              Edit Category
+              Add Category
             </span>
             {/* Close Modal */}
             <span
               style={{ background: '#626262' }}
-              onClick={() => {
-                console.log('background clicked')
-                dispatch(editCategoryModalOpen({ open: false }))
-              }}
+              onClick={(e) => dispatch(addCategoryModal(false))}
               className="cursor-pointer text-gray-100 py-2 px-2 rounded-full"
             >
               <svg
@@ -237,10 +229,10 @@ const EditCategoryModal = () => {
                 <label htmlFor="name">Category Name </label>
                 <span className="text-red-600 text-xs">* Required</span>
                 <input
-                  value={editFormData.category_name}
+                  value={fData.category_name}
                   onChange={(e) =>
-                    setEditformdata({
-                      ...editFormData,
+                    setFdata({
+                      ...fData,
                       error: false,
                       success: false,
                       category_name: e.target.value,
@@ -255,10 +247,10 @@ const EditCategoryModal = () => {
                 <label htmlFor="status">Category Status </label>
                 <span className="text-red-600 text-xs">* Required</span>
                 <select
-                  value={editFormData.category_status}
+                  value={fData.category_status}
                   onChange={(e) =>
-                    setEditformdata({
-                      ...editFormData,
+                    setFdata({
+                      ...fData,
                       error: false,
                       success: false,
                       category_status: e.target.value,
@@ -288,10 +280,10 @@ const EditCategoryModal = () => {
                 </span>
                 {imageAdded && (
                   <div className="mt-3">
-                    {editFormData.category_image && (
+                    {fData.category_image && (
                       <div className="relative">
                         <img
-                          src={editFormData.category_image}
+                          src={fData.category_image}
                           alt="Category_Image"
                           className="w-20 h-20 object-cover rounded-full mr-2 mb-2"
                         />
@@ -299,7 +291,7 @@ const EditCategoryModal = () => {
                           type="button"
                           className="absolute top-0 right-0 p-1 text-gray-500 hover:text-gray-700 transition-all duration-150"
                           onClick={(e) =>
-                            handleImageRemove(editFormData.category_image, e)
+                            handleImageRemove(fData.category_image, e)
                           }
                         >
                           &times;
@@ -342,10 +334,10 @@ const EditCategoryModal = () => {
                 <label htmlFor="description">Category Description</label>
                 <span className="text-red-600 text-xs">* Required</span>
                 <textarea
-                  value={editFormData.category_description}
+                  value={fData.category_description}
                   onChange={(e) =>
-                    setEditformdata({
-                      ...editFormData,
+                    setFdata({
+                      ...fData,
                       error: false,
                       success: false,
                       category_description: e.target.value,
@@ -368,9 +360,9 @@ const EditCategoryModal = () => {
                 style={{
                   backgroundColor: '#626262',
                 }}
-                className="rounded-xl  w-40 text-gray-100 text-lg font-medium py-2"
+                className="rounded-xl  w-32 text-gray-100 text-lg font-medium py-2"
               >
-                Update Category
+                Add Category
               </button>
             </div>
           </form>
@@ -380,4 +372,4 @@ const EditCategoryModal = () => {
   )
 }
 
-export default EditCategoryModal
+export default FFAddCategoryModal
