+++ src/components/molecules/FFModal/FFEditProductModal.jsx
@@ -1,31 +1,36 @@
-import React, { useState, useEffect, useRef } from 'react'
-import { imageUpload, removeImage } from '../../utils/imagesFunctions'
-import { useDispatch, useSelector } from 'react-redux'
+import React, { Fragment, useState, useEffect, useRef } from 'react'
+import { useSelector, useDispatch } from 'react-redux'
 import {
-  add_product_modal,
-  postProduct,
+  edit_product_modal,
   getAllProducts,
   isSuccess,
   isError,
-  addProductModal,
+  editProductModalOpen,
+  updateProduct,
   message,
 } from '../../redux/features/productSlice'
-
 import { selectAllCategories } from '../../redux/features/categorySlice'
+import { imageUpload, removeImage } from '../../utils/imagesFunctions'
 import {
-  validateNumericInput,
   validateProductName,
+  validateNumericInput,
 } from '../../utils/validations/product'
-const AddProductDetail = () => {
+const FFEditProductModal = () => {
   const dispatch = useDispatch()
-  const add_product_detail_modal = useSelector(add_product_modal)
+  const edit_product_detail_modal = useSelector(edit_product_modal)
   const succesStatus = useSelector(isSuccess)
   const errorStatus = useSelector(isError)
   const responseMessage = useSelector(message)
+
+  const [selectedFile, setSelectedFile] = useState(null)
+  const [error, setError] = useState('')
+  const fileInputRef = useRef(null)
+
+  const categories = useSelector(selectAllCategories)
+  const [imageAdded, setImageAdded] = useState(false)
   //error
   const [productNameError, setProductNameError] = useState('')
 
-  const categories = useSelector(selectAllCategories)
   const alert = (msg, type) => (
     <div
       className={`text-${type} flex felx-row  justify-center items-center py-2 px-4 w-full`}
@@ -34,17 +39,10 @@ const AddProductDetail = () => {
     </div>
   )
 
-  //added newly
-  const [selectedFile, setSelectedFile] = useState(null)
-  const [error, setError] = useState('')
-  const fileInputRef = useRef(null)
-
-  const [imageAdded, setImageAdded] = useState(false)
-
-  const [fData, setFdata] = useState({
+  const [editformData, setEditformdata] = useState({
     product_name: '',
     product_price: 0.0,
-    product_status: 'in stock',
+    product_status: '',
     product_category: '',
     product_offer: 0.0,
     product_images: [],
@@ -57,69 +55,63 @@ const AddProductDetail = () => {
   })
 
   useEffect(() => {
+    setEditformdata({
+      product_name: edit_product_detail_modal.product_name,
+      product_price: edit_product_detail_modal.product_price,
+      product_status: edit_product_detail_modal.product_status,
+      product_category: edit_product_detail_modal.product_category,
+      product_offer: edit_product_detail_modal.product_offer,
+      product_images: edit_product_detail_modal.product_images,
+      product_quantity: edit_product_detail_modal.product_quantity,
+      product_visible: edit_product_detail_modal.product_visible,
+      product_weight: edit_product_detail_modal.product_weight,
+      product_sale_status: edit_product_detail_modal.product_sale_status,
+    })
+  }, [edit_product_detail_modal])
+
+  useEffect(() => {
     setImageAdded(true)
-    setSelectedFile(null)
-  }, [fData?.product_images])
+  }, [editformData?.product_images])
 
-  const handleSubmit = async (e) => {
+  const submitForm = async (e) => {
     e.preventDefault()
-
-    if (fData.product_images?.length > 1) {
-      setFdata({ ...fData, error: 'Please upload at least 1 image' })
-      setTimeout(() => {
-        setFdata({ ...fData, error: false })
-      }, 2000)
+    if (editformData?.product_images > 1) {
+      console.log('UploaImage ')
+    } else {
+      console.log('Image uploading')
     }
-
     try {
-      const { success, error, ...data } = fData
-      dispatch(postProduct(data))
-
+      const { error, success, ...updatedData } = editformData
+      console.log(editformData)
+      dispatch(
+        updateProduct({
+          ...updatedData,
+          product_id: edit_product_detail_modal.product_id,
+        }),
+      )
       if (succesStatus) {
         dispatch(getAllProducts())
-        setFdata({
-          ...fData,
-          product_name: '',
-          product_price: 0.0,
-          product_status: '',
-          product_category: '',
-          product_offer: 0.0,
-          product_images: [],
-          product_quantity: 0,
-          product_visible: true,
-          product_weight: 0.0,
-          product_sale_status: false,
-          success: succesStatus,
-          error: false,
-        })
+        setEditformdata({ ...editformData, success: succesStatus })
         setTimeout(() => {
-          setFdata({
-            ...fData,
-            product_name: '',
-            product_price: 0.0,
-            product_status: '',
-            product_category: '',
-            product_offer: 0.0,
-            product_images: [],
-            product_quantity: 0,
-            product_visible: true,
-            product_weight: 0.0,
-            product_sale_status: false,
-            success: false,
-            error: false,
+          return setEditformdata({
+            ...editformData,
+            success: succesStatus,
           })
         }, 2000)
       } else if (errorStatus) {
-        setFdata({ ...fData, success: false, error: errorStatus })
+        setEditformdata({ ...editformData, error: errorStatus })
         setTimeout(() => {
-          return setFdata({ ...fData, error: false, success: false })
+          return setEditformdata({
+            ...editformData,
+            error: false,
+            success: false,
+          })
         }, 2000)
       }
     } catch (error) {
       console.log(error)
     }
   }
-
   const handleFileSelect = (event) => {
     const file = event.target.files[0]
     const allowedTypes = ['image/jpeg', 'image/png', 'image/gif']
@@ -143,10 +135,10 @@ const AddProductDetail = () => {
     if (!selectedFile) {
       return
     }
-    imageUpload(selectedFile, 'itemsImages')
+    imageUpload(selectedFile, 'productImages')
       .then((imageUrl) => {
         // push the imageUrl to the imageUrl array
-        setFdata((prevState) => ({
+        setEditformdata((prevState) => ({
           ...prevState,
           error: false,
           success: false,
@@ -166,7 +158,7 @@ const AddProductDetail = () => {
     removeImage(imageUrl)
       .then(
         // Remove the imageUrl from the formData
-        setFdata((prevState) => ({
+        setEditformdata((prevState) => ({
           ...prevState,
           error: false,
           success: false,
@@ -185,13 +177,11 @@ const AddProductDetail = () => {
 
   return (
     <>
+      {/* Black Overlay */}
       <div
-        onClick={() => {
-          console.log('background clicked')
-          dispatch(addProductModal(false))
-        }}
+        onClick={(e) => dispatch(editProductModalOpen({ open: false }))}
         className={`${
-          add_product_detail_modal ? '' : 'hidden'
+          edit_product_detail_modal.modal ? '' : 'hidden'
         } fixed top-0 left-0 z-30 w-full h-full bg-black opacity-50`}
       />
       {/* End Black Overlay */}
@@ -199,18 +189,18 @@ const AddProductDetail = () => {
       {/* Modal Start */}
       <div
         className={`${
-          add_product_detail_modal ? '' : 'hidden'
+          edit_product_detail_modal.modal ? '' : 'hidden'
         } fixed inset-0 flex items-center z-30 justify-center overflow-auto`}
       >
         <div className="mt-4 md:mt-0 relative bg-[#F5F3F0] w-8/12 md:w-3/6 shadow-lg flex flex-col items-center space-y-4 px-4 py-4 md:px-8 rounded-xl">
           <div className="flex items-center justify-between w-full pt-4">
             <span className="text-left font-semibold text-2xl tracking-wider">
-              Add Product
+              Edit Product
             </span>
             {/* Close Modal */}
             <span
               style={{ background: '#626262' }}
-              onClick={(e) => dispatch(addProductModal(false))}
+              onClick={(e) => dispatch(editProductModalOpen({ open: false }))}
               className="cursor-pointer text-gray-100 py-2 px-2 rounded-full"
             >
               <svg
@@ -229,39 +219,39 @@ const AddProductDetail = () => {
               </svg>
             </span>
           </div>
-          {errorStatus
+          {editformData.error
             ? alert(
                 responseMessage !== 'All products' ? responseMessage : '',
                 'red',
               )
             : ''}
-          {succesStatus
+          {editformData.success
             ? alert(
                 responseMessage !== 'All products' ? responseMessage : '',
                 'green',
               )
             : ''}
           {productNameError ? alert(productNameError, 'red') : ''}
-          <form className="w-full">
+          <form className="w-full" onSubmit={(e) => submitForm(e)}>
             <div className="flex space-x-1 py-4">
               <div className="w-1/2 flex flex-col space-y-1 space-x-1">
                 <label htmlFor="name">Product Name </label>
                 <span className="text-red-600 text-xs">* Required</span>
                 <input
-                  value={fData.product_name}
+                  value={editformData.product_name}
                   onChange={(e) => {
                     validateProductName({
                       productName: e.target.value,
                       setProductNameError,
                     })
-                    setFdata({
-                      ...fData,
+                    setEditformdata({
+                      ...editformData,
                       error: false,
                       success: false,
                       product_name: e.target.value,
                     })
                   }}
-                  className="px-4 py-2 border h-14 rounded-lg  focus:outline-none"
+                  className="px-4 py-2 border  h-14 focus:outline-none"
                   type="text"
                 />
               </div>
@@ -269,24 +259,24 @@ const AddProductDetail = () => {
                 <label htmlFor="price">Product Price (Rs.) </label>
                 <span className="text-red-600 text-xs">* Required</span>
                 <input
-                  value={fData.product_price}
+                  value={editformData.product_price}
                   onChange={(e) =>
-                    setFdata({
-                      ...fData,
+                    setEditformdata({
+                      ...editformData,
                       error: false,
                       success: false,
                       product_price: e.target.value,
                     })
                   }
                   onKeyDown={(e) => validateNumericInput(e)}
-                  className="px-4 py-2 border h-14 rounded-lg  focus:outline-none"
+                  type="number"
+                  className="px-4 py-2 border  h-14 focus:outline-none"
                   id="price"
                 />
               </div>
             </div>
 
             {/* Most Important part for uploading multiple image */}
-
             <div className="flex space-x-1 py-4">
               <div className="flex w-1/2  flex-col mt-4">
                 <label htmlFor="image">Image Upload</label>
@@ -295,8 +285,8 @@ const AddProductDetail = () => {
                 </span>
                 {imageAdded && (
                   <div className="mt-3 flex flex-row space-x-1">
-                    {fData?.product_images &&
-                      fData?.product_images.map((image, index) => (
+                    {editformData?.product_images &&
+                      editformData?.product_images.map((image, index) => (
                         <div className="relative " key={index}>
                           <img
                             src={image}
@@ -347,39 +337,40 @@ const AddProductDetail = () => {
                 <label htmlFor="weight">Product Weight (Kg)</label>
                 <span className="text-red-600 text-xs">* Required</span>
                 <input
-                  value={fData.product_weight}
+                  value={editformData.product_weight}
+                  type="number"
                   onChange={(e) =>
-                    setFdata({
-                      ...fData,
+                    setEditformdata({
+                      ...editformData,
                       error: false,
                       success: false,
                       product_weight: e.target.value,
                     })
                   }
                   onKeyDown={(e) => validateNumericInput(e)}
-                  type="number"
-                  className="px-4 py-2 border h-14 rounded-lg focus:outline-none "
+                  className="px-4 py-2   h-14 border focus:outline-none"
                   id="weight"
                 />
               </div>
             </div>
+
             {/* Most Important part for uploading multiple image */}
             <div className="flex space-x-1 py-4">
               <div className="w-1/2 flex flex-col space-y-1">
                 <label htmlFor="status">Product Status </label>
                 <span className="text-red-600 text-xs">* Required</span>
                 <select
-                  value={fData.product_status}
+                  value={editformData.product_status}
                   onChange={(e) =>
-                    setFdata({
-                      ...fData,
+                    setEditformdata({
+                      ...editformData,
                       error: false,
                       success: false,
                       product_status: e.target.value,
                     })
                   }
                   name="status"
-                  className="px-4 py-2 border h-14 rounded-lg  focus:outline-none"
+                  className="px-4 py-2  h-14 border focus:outline-none"
                   id="status"
                 >
                   <option name="status" value="in stock">
@@ -394,31 +385,35 @@ const AddProductDetail = () => {
                 <label htmlFor="status">Product Category </label>
                 <span className="text-red-600 text-xs">* Required</span>
                 <select
-                  value={fData.product_category}
                   onChange={(e) =>
-                    setFdata({
-                      ...fData,
+                    setEditformdata({
+                      ...editformData,
                       error: false,
                       success: false,
                       product_category: e.target.value,
                     })
                   }
                   name="status"
-                  className="px-4 py-2 border h-14 focus:outline-none"
+                  className="px-4 py-2 border  h-14 focus:outline-none"
                   id="status"
                 >
                   <option disabled value="">
                     Select a category
                   </option>
-                  {categories.length > 0
-                    ? categories.map(function (elem) {
-                        return (
-                          <option name="status" value={elem._id} key={elem._id}>
-                            {elem.category_name}
-                          </option>
-                        )
-                      })
-                    : ''}
+                  {categories.map((elem) => {
+                    return (
+                      <Fragment key={elem._id}>
+                        <option
+                          name="status"
+                          value={elem._id}
+                          key={elem._id}
+                          selected
+                        >
+                          {elem.category_name}
+                        </option>
+                      </Fragment>
+                    )
+                  })}
                 </select>
               </div>
             </div>
@@ -427,10 +422,10 @@ const AddProductDetail = () => {
                 <label htmlFor="quantity">Product in Stock </label>
                 <span className="text-red-600 text-xs">* Required</span>
                 <input
-                  value={fData.product_quantity}
+                  value={editformData.product_quantity}
                   onChange={(e) =>
-                    setFdata({
-                      ...fData,
+                    setEditformdata({
+                      ...editformData,
                       error: false,
                       success: false,
                       product_quantity: e.target.value,
@@ -438,7 +433,7 @@ const AddProductDetail = () => {
                   }
                   onKeyDown={(e) => validateNumericInput(e)}
                   type="number"
-                  className="px-4 py-2 border h-14 rounded-lg  focus:outline-none"
+                  className="px-4 py-2 border  h-14 focus:outline-none"
                   id="quantity"
                 />
               </div>
@@ -446,33 +441,29 @@ const AddProductDetail = () => {
                 <label htmlFor="offer">Product Offfer (%) </label>
                 <span className="text-red-600 text-xs">* Required</span>
                 <input
-                  value={fData.product_offer}
-                  onKeyDown={(e) => validateNumericInput(e)}
-                  type="number"
+                  value={editformData.product_offer}
                   onChange={(e) =>
-                    setFdata({
-                      ...fData,
+                    setEditformdata({
+                      ...editformData,
                       error: false,
                       success: false,
                       product_offer: e.target.value,
                     })
                   }
-                  className="px-4 py-2 border  h-14 rounded-lg focus:outline-none"
+                  type="number"
+                  onKeyDown={(e) => validateNumericInput(e)}
+                  className="px-4 py-2 border h-14 focus:outline-none"
                   id="offer"
                 />
               </div>
             </div>
             <div className="flex flex-row space-y-1 w-full pb-4 md:pb-6 mt-4 justify-end">
-              {/* button needs to be replaced */}
               <button
-                type="button"
-                onClick={(e) => handleSubmit(e)}
-                style={{
-                  backgroundColor: '#626262',
-                }}
-                className="rounded-xl  w-32 text-gray-100 text-lg font-medium py-2"
+                style={{ background: '#626262' }}
+                type="submit"
+                className="rounded-xl  w-36 text-gray-100 text-lg font-medium py-2"
               >
-                Add product
+                Update product
               </button>
             </div>
           </form>
@@ -482,4 +473,4 @@ const AddProductDetail = () => {
   )
 }
 
-export default AddProductDetail
+export default FFEditProductModal
