+++ src/pages/category/AdminCategoryPage.jsx
@@ -15,13 +15,13 @@ import FFTableHead from '../../components/molecules/FFTable/FFTableHead/FFTableH
 import FFButton from '../../components/atoms/FFButton/FFButton'
 import FFTable from '../../components/molecules/FFTable/FFTable'
 import FFTableBody from '../../components/molecules/FFTable/FFTableBody/FFTableBody'
-import AlertModal from '../../components/organisms/AlertModal'
+import FFAlertModal from '../../components/organisms/FFAlertModal'
 
 import FFCategoryTableData from '../../components/atoms/FFCategoryTableData/FFCategoryTableData'
 import { CATEGORY_TABLE_HEADER_CONTENT } from '../../utils/constants'
-import AddCategoryModal from '../../components/organisms/AddCategoryModal'
-import EditCategoryModal from '../../components/organisms/EditCategoryModal'
-import { ViewCategoryModal } from '../../components/organisms/ViewCategoryModal'
+import FFAddCategoryModal from '../../components/molecules/FFModal/FFAddCategoryModal'
+import FFEditCategoryModal from '../../components/organisms/FFEditCategoryModal'
+import { FFViewCategoryModal } from '../../components/organisms/FFViewCategoryModal'
 const AdminCategoryPage = () => {
   const dispatch = useDispatch()
   const categoryList = useSelector(selectAllCategories)
@@ -83,10 +83,10 @@ const AdminCategoryPage = () => {
           },
         }}
       />
-      <AddCategoryModal />
-      <AlertModal />
-      <EditCategoryModal />
-      <ViewCategoryModal />
+      <FFAddCategoryModal />
+      <FFAlertModal />
+      <FFEditCategoryModal />
+      <FFViewCategoryModal />
       {content}
     </div>
   )
